Add skeleton for landing page and home
This adds 2 different roots for logged in and not logged in.
I think this'll be nice to have as a constraint at the routing level, so
that we can just ignore this and not worry about it in the code itself.
This is all shit that's tested in rails and realistically once I get
past this I won't need to think about it much......